seperate evil engine into another file
